@@ -1,6 +1,9 @@
 using System;
+using System.Threading;
 using System.Windows.Forms;
+using Chat.UI.Authorization;
 using Infrastructure;
+using Infrastructure.Exceptions;
 
 namespace Chat.UI
 {
@@ -12,10 +15,10 @@ namespace Chat.UI
         [STAThread]
         static void Main()
         {
-            System.Windows.Forms.Application.SetHighDpiMode(HighDpiMode.SystemAware);
-            System.Windows.Forms.Application.EnableVisualStyles();
-            System.Windows.Forms.Application.SetCompatibleTextRenderingDefault(false);
-            System.Windows.Forms.Application.Run(new AuthorizeForm());
+            Application.SetHighDpiMode(HighDpiMode.SystemAware);
+            Application.EnableVisualStyles();
+            Application.SetCompatibleTextRenderingDefault(false);
+            Application.Run(new AuthorizeForm());
         }
     }
 }
