@@ -9,7 +9,14 @@ public static class RoomsViewModelExtensions
     public static RoomsViewModel ToHubsViewModel(this IEnumerable<ChatRoom> chatRooms)
     {
         var hubsModels = chatRooms
-            .Select(x => new RoomViewModel {ActiveUsers = x.Users.Count, Id = x.Id});
+            .Select(x => new RoomViewModel
+            {
+                ActiveUsers = x.UsersCount, 
+                Id = x.Id,
+                Name = x.Name,
+                Protection = x.RequiresPassword,
+                MaxCapacity = x.MaxUsers
+            });
         return new RoomsViewModel {ChatRooms = hubsModels.ToArray()};
     }
 }
@@ -23,4 +30,7 @@ public class RoomViewModel
 {
     public int ActiveUsers;
     public Guid Id;
+    public string Name;
+    public bool Protection;
+    public int MaxCapacity;
 }
